GUI Updates
- Create Plots (for multi Runs)
- list with multiRuns tests
- Start, rate button for multi run

+++ src/source/DataGenerator.java
@@ -148,7 +148,7 @@ public class DataGenerator {
 		cars.get(car).setEndTime(new_end);
 	}
 
-	public ArrayList<Car> getCars(int count) {
+	public ArrayList<Car> getCars(double count) {
 		
 		if(count == cars.size() || (count == -1))
 		{
